[EST-35]Refactor: Service 클래스 리팩토링
