+++ src/main/java/com/blackshoe/esthete/exception/TokenErrorResult.java
@@ -1,17 +1,17 @@
-package com.blackshoe.esthete.exception;
-
-public enum TokenErrorResult {
-    TOKEN_EXPIRED("Token expired"),
-    INVALID_TOKEN("Invalid token"),
-    ;
-
-    private final String message;
-
-    TokenErrorResult(String message) {
-        this.message = message;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-}
+//package com.blackshoe.esthete.exception;
+//
+//public enum TokenErrorResult {
+//    TOKEN_EXPIRED("Token expired"),
+//    INVALID_TOKEN("Invalid token"),
+//    ;
+//
+//    private final String message;
+//
+//    TokenErrorResult(String message) {
+//        this.message = message;
+//    }
+//
+//    public String getMessage() {
+//        return message;
+//    }
+//}
