+++ src/main/java/com/blackshoe/esthete/exception/TokenException.java
@@ -1,14 +1,14 @@
-package com.blackshoe.esthete.exception;
-
-public class TokenException extends RuntimeException {
-
-    private final TokenErrorResult errorResult;
-
-    public TokenException(TokenErrorResult errorResult) {
-        this.errorResult = errorResult;
-    }
-
-    public TokenErrorResult getErrorResult() {
-        return errorResult;
-    }
-}
+//package com.blackshoe.esthete.exception;
+//
+//public class TokenException extends RuntimeException {
+//
+//    private final TokenErrorResult errorResult;
+//
+//    public TokenException(TokenErrorResult errorResult) {
+//        this.errorResult = errorResult;
+//    }
+//
+//    public TokenErrorResult getErrorResult() {
+//        return errorResult;
+//    }
+//}
