+++ src/main/java/com/blackshoe/esthete/service/RedisService.java
@@ -1,0 +1,10 @@
+package com.blackshoe.esthete.service;
+
+public interface RedisService {
+    String getData(String key);
+    void setData(String key,String value);
+    boolean existsKey(String key);
+    void setDataExpire(String key,String value,long duration);
+    void deleteData(String key);
+
+}
